fix(e2e): stabilize project deletion and workflow creation tests
- updated deleteProjectViaUI to handle type-to-confirm dialogs
- added centralized createWorkflowViaBuilder helper to handle 4-step wizard
- refactored workflows and deep_check specs to use new helpers
- added robust input filling with retries and step waits
